--- qudini-reactive-utils/src/main/java/com/qudini/reactive/utils/intervals/temporal/OffsetDateTimeInterval.java
@@ -1,31 +1,0 @@
-package com.qudini.reactive.utils.intervals.temporal;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
-import lombok.extern.jackson.Jacksonized;
-
-import java.time.OffsetDateTime;
-import java.time.temporal.TemporalUnit;
-
-@Value
-@Builder
-@Jacksonized
-@AllArgsConstructor(staticName = "of")
-public class OffsetDateTimeInterval implements TemporalInterval<OffsetDateTime, OffsetDateTimeInterval> {
-
-    OffsetDateTime start;
-    OffsetDateTime end;
-
-    @Override
-    public @NonNull OffsetDateTimeInterval withStart(OffsetDateTime start) {
-        return new OffsetDateTimeInterval(start, end);
-    }
-
-    @Override
-    public @NonNull OffsetDateTimeInterval withEnd(OffsetDateTime end) {
-        return new OffsetDateTimeInterval(start, end);
-    }
-
-}
