--- qudini-reactive-utils/src/main/java/com/qudini/reactive/utils/intervals/temporal/OffsetTimeInterval.java
@@ -1,31 +1,0 @@
-package com.qudini.reactive.utils.intervals.temporal;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
-import lombok.extern.jackson.Jacksonized;
-
-import java.time.OffsetTime;
-import java.time.temporal.TemporalUnit;
-
-@Value
-@Builder
-@Jacksonized
-@AllArgsConstructor(staticName = "of")
-public class OffsetTimeInterval implements TemporalInterval<OffsetTime, OffsetTimeInterval> {
-
-    OffsetTime start;
-    OffsetTime end;
-
-    @Override
-    public @NonNull OffsetTimeInterval withStart(OffsetTime start) {
-        return new OffsetTimeInterval(start, end);
-    }
-
-    @Override
-    public @NonNull OffsetTimeInterval withEnd(OffsetTime end) {
-        return new OffsetTimeInterval(start, end);
-    }
-
-}
